move mock to constants
